+++ src/aixd_ara/components/ara_PlotCorrelations/code.py
@@ -10,13 +10,7 @@ cid = component_id(session_id(), ghenv.Component, "PlotCorrelations")
 
 
 if plot:
-    st[cid] = plot_correlations(
-        session_id(), variables, output_type
-    )  # if output_type interactive: will launch the plotly fig in browser
+    st[cid] = plot_correlations(session_id(), variables, "interactive")  # will launch the plotly fig in browser
 
 if cid in st.keys():
     print(st[cid])
-    # TODO: add error msg here
-    if output_type == "static" and "imgstr" in st[cid].keys():
-        imgstr = st[cid]["imgstr"]
-        img = convert_str_to_bitmap(imgstr, scale)