+++ CHANGELOG.md
@@ -10,13 +10,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 ### Changed
+<<<<<<< 46-update-examples
 
 * Updated examples & documentation
+=======
+* Fixed domain init in DataInt
+* Changed output options in Plot components.
+>>>>>>> main
 * Fixed Windows install script trying to install from offline wheels.
 * Fixed issues related to using ARA on Mac.
 
 ### Removed
-
+* Removed option for static plots.
 
 ## [0.9.3] 2024-06-18